move LoginUserResponse class to correct package (domain)
