+++ app/src/main/java/com/example/tasky/feature_login/domain/model/LoginUserResponse.kt
@@ -1,0 +1,9 @@
+package com.example.tasky.feature_login.domain.model
+
+data class LoginUserResponse(
+    val accessToken: String,
+    val refreshToken: String,
+    val fullName: String,
+    val userId: String,
+    val accessTokenExpirationTimestamp: Long
+)
